Refactoring: cpp unit
Вынесла часть для плюсов
