--- task_2/classunit.cpp
@@ -1,35 +1,0 @@
-#include "classunit.h"
-
-#include <sstream>
-
-void CppClassUnit::add(const std::shared_ptr<Unit> &unit, Unit::Flags flags)
-{
-    auto accessModifier = PRIVATE;
-    if(flags < ACCESS_MODIFIERS.size()) {
-        accessModifier = static_cast<AccessModifier>(flags);
-    }
-    m_fields[accessModifier].push_back(unit);
-}
-
-std::string CppClassUnit::compile(unsigned int level) const
-{
-    std::stringstream result;
-    result << core::generateShift(level)
-           << "class "
-           << m_name
-           << " {\n";
-
-    for(size_t i = 0; i < ACCESS_MODIFIERS.size(); ++i) {
-        if(m_fields[i].empty()) {
-            continue;
-        }
-        result << ACCESS_MODIFIERS[i] << ":\n";
-        for(const auto& f : m_fields[i]) {
-            result << f->compile(level + 1);
-        }
-        result << "\n";
-    }
-    result << core::generateShift(level) << "};\n";
-
-    return result.str();
-}
