--- task_2/printoperatorunit.cpp
@@ -1,13 +1,0 @@
-#include "printoperatorunit.h"
-
-#include <sstream>
-
-std::string CppPrintOperatorUnit::compile(unsigned int level) const
-{
-    std::stringstream ss;
-    ss << core::generateShift(level)
-       << "printf( \""
-       << m_text
-       << "\" );\n";
-    return ss.str();
-}
