--- task_2/printoperatorunit.h
@@ -1,12 +1,0 @@
-#ifndef PRINTOPERATORUNIT_H
-#define PRINTOPERATORUNIT_H
-
-#include "unit.h"
-
-class CppPrintOperatorUnit : public core::PrintOperatorUnit {
-public:
-    using PrintOperatorUnit::PrintOperatorUnit;
-    std::string compile( unsigned int level = 0 ) const;
-};
-
-#endif // PRINTOPERATORUNIT_H
